Safari-609.1.11.1/LayoutTests/svg/dom/svglist-insertItemBefore-appends-expected.txt

Tests that insertItemBefore correctly appends if its index is out of bounds.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS svgList.numberOfItems is 0
PASS svgList.numberOfItems is 1
PASS svgList.getItem(0) is seg01
PASS svgList.numberOfItems is 2
PASS svgList.getItem(0) is seg01
PASS svgList.getItem(1) is seg11
PASS svgList.numberOfItems is 3
PASS svgList.getItem(0) is seg00
PASS svgList.getItem(1) is seg01
PASS svgList.getItem(2) is seg11
PASS svgList.numberOfItems is 4
PASS svgList.getItem(0) is seg00
PASS svgList.getItem(1) is seg01
PASS svgList.getItem(2) is seg11
PASS svgList.getItem(3) is seg10
PASS successfullyParsed is true

TEST COMPLETE
